fix linters and separate DOM from logic
